Use a bogus tag instead of a list for marks
This replaces the internal list of message-ids with the regular notmuch
tagging mechanism, and (too?) clever use of notmuch's customization to
make the mark more visible.

A marked message contains the "ohm_bogus" tag, so the (useful) tagging
operations can use "tag:ohm_bogus" as the search query afterwards.

Toggling is done via `evil-collection-notmuch`, though it's probably
easy to reimplement this and avoid the extra dependency.

Marks are not removed automatically.

In tree buffers, the advice re-applies text properties to include the
mark's face. In search buffers, however, it is up to the user to define
`notmuch-search-line-faces` and `notmuch-tag-formats` accordingly. For
example:

  (setq notmuch-tag-formats
        '(
          ;;others...
          ("ohm_bogus" nil) ; hides the mark from the tag list
          ))

  (setq notmuch-search-line-faces
        '(("unread" . notmuch-search-unread-face)
          ("flagged" . notmuch-search-flagged-face)
          ;;others...
          ("ohm_bogus" . (:underline (:style wave :color "darkgray"))) ;a squiggly line under the message
          ))